requirements.txt
ollama
sqlalchemy
pandas
matplotlib
python-dotenv
